Moved ConsentStringBuilder-related classed to 'utils'-package as @chrispaterson suggested.
